+++ js/ps.js
@@ -6,14 +6,25 @@ pageStats.createEvent = function(name, value){
     }
 }
 
-console.log('PageStats::ping', {interval: pageStats.config.ping*1000})
+pageStats.pingCount = 0;
+
+/**
+ * incremental time on page metrics.
+ * first pageStats.config.ping seconds send ping once per second, then only
+ * every pageStats.config.ping seconds
+ */
 function time_on_page() {
     fetch(pageStats.url, {
         method: 'POST',
         body: JSON.stringify(pageStats.createEvent('ping', true))
     });
+
+    pageStats.pingCount++;
+
+    interval = pageStats.pingCount < pageStats.config.ping ? 1 : pageStats.config.ping;
+    window.setTimeout(time_on_page, interval * 1000);
+
 }
-window.setInterval(time_on_page, pageStats.config.ping * 1000);
 
 //collect stats on page load
 time_on_page()